Big commit here. Adds all pages and css to site.
There's still a lot of work to be done here. First I need more images,
second I need to rewrite / organize the content to flow better based on
feedback, and lastly I need to add more content to the news feed and
more. A separate registration section might be more appropriate than on
the front page where it is now.